wip: Deduplicates code
